+++ packages/message-bar/__tests__/shared.base.js
@@ -1,0 +1,9 @@
+import { iterator } from "@times-components/test-utils";
+import MessageBarTests from "./message-bar.base";
+import MessageManagerTests from "./message-manager.base";
+
+export default animate => {
+  const tests = [...MessageBarTests(animate), ...MessageManagerTests(animate)];
+
+  iterator(tests);
+};