+++ packages/message-bar/src/styles/index.js
@@ -1,0 +1,6 @@
+import { StyleSheet } from "react-native";
+import sharedStyles from "./shared";
+
+const styles = scale => StyleSheet.create(sharedStyles(scale));
+
+export default styles;